Merge

+++ src/main/java/ar/edu/unq/desapp/grupoI/backenddesappapi/dto/ReviewDTO.java
@@ -3,7 +3,6 @@ package ar.edu.unq.desapp.grupoi.backenddesappapi.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.time.LocalDate;
 import java.util.Date;
 
 @Getter
@@ -12,7 +11,7 @@ public class ReviewDTO {
     public String tittle_tconst;
     public String resume;
     public String extendedText;
-    public float rating;
+    public long rating;
     public Date date;
     public String language;
     public Long userId;
